requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
chromadb
pypdf
python-dotenv
tiktoken
